+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v0.3.0dev0 [XXX-XX-XX]
 
+### Fixed
+
+- [#16](https://github.com/Clinical-Genomics/paraphrase/pull/16) - Paraphase 4.x
+  `region_specific_info` fields now use the existing flat output and rules
+  contract
+
 ### Updated
 
 - [#14](https://github.com/Clinical-Genomics/paraphrase/pull/14) - Updated testdata to paraphase 3.5.0